src/processors/text_processor/requirements.txt
lxml==5.3.1
pyyaml==6.0.2
striprtf==0.0.28